+++ src/c/utils/clock-posix.cc
@@ -6,5 +6,5 @@
 uint64_t SystemRealTimeClock::millis_since_epoch_utc() {
   struct timespec spec;
   clock_gettime(CLOCK_REALTIME, &spec);
-  return static_cast<uint64_t>((spec.tv_sec * 1000.0) + (spec.tv_nsec / 1000000.0));
+  return static_cast<uint64_t>((static_cast<double>(spec.tv_sec) * 1000.0) + (static_cast<double>(spec.tv_nsec) / 1000000.0));
 }
