+++ src/c/utils/crash-execinfo.c
@@ -6,7 +6,7 @@
 #include <execinfo.h>
 #include <unistd.h>
 
-static const size_t kMaxStackSize = 128;
+static const int kMaxStackSize = 128;
 
 void initialize_crash_handler() {
   // nothing to initialize
@@ -15,9 +15,9 @@ void initialize_crash_handler() {
 void print_stack_trace(out_stream_t *out, int signum) {
   out_stream_printf(out, "# Received condition %i\n", signum);
   void *raw_frames[kMaxStackSize];
-  size_t size = backtrace(raw_frames, kMaxStackSize);
+  int size = backtrace(raw_frames, kMaxStackSize);
   char **frames = backtrace_symbols(raw_frames, size);
-  for (size_t i = 0; i < size; i++) {
+  for (int i = 0; i < size; i++) {
     out_stream_printf(out, "# - %s\n", frames[i]);
   }
   free(frames);
